+++ lib/mix/tasks/ua_inspector/download.databases.ex
@@ -13,19 +13,9 @@ defmodule Mix.Tasks.UaInspector.Download.Databases do
 
   alias Mix.UAInspector.Download
   alias UAInspector.Config
-  alias UAInspector.Database
   alias UAInspector.Downloader
 
   use Mix.Task
-
-  @databases [
-    Database.Bots,
-    Database.BrowserEngines,
-    Database.Clients,
-    Database.Devices,
-    Database.OSs,
-    Database.VendorFragments
-  ]
 
   def run(args) do
     Mix.shell().info("UAInspector Database Download")
@@ -44,25 +34,10 @@ defmodule Mix.Tasks.UaInspector.Download.Databases do
 
   defp run_confirmed(true) do
     {:ok, _} = Application.ensure_all_started(:hackney)
-    :ok = Download.prepare_database_path()
-    :ok = download(@databases)
+    :ok = Downloader.download(:databases)
 
     Mix.shell().info("Download complete!")
 
     :ok
-  end
-
-  defp download([]), do: :ok
-
-  defp download([database | databases]) do
-    for {_type, local, remote} <- database.sources do
-      Mix.shell().info(".. downloading: #{local}")
-
-      target = Path.join([Config.database_path(), local])
-
-      Downloader.download_file(remote, target)
-    end
-
-    download(databases)
   end
 end