+++ verify/lib/mix/tasks/ua_inspector/verify.ex
@@ -57,8 +57,7 @@ defmodule Mix.Tasks.UaInspector.Verify do
   defp maybe_download(quick: true), do: :ok
 
   defp maybe_download(_) do
-    :ok = Mix.Tasks.UaInspector.Download.Databases.run(["--force"])
-    :ok = Mix.Tasks.UaInspector.Download.ShortCodeMaps.run(["--force"])
+    :ok = Downloader.download()
     :ok = Verify.Fixtures.download()
 
     Mix.shell().info("=== Skip downloads using '--quick' ===")