+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Enhancements
     - All databases can be reloaded (asynchronously) using
-      `UAInspector.reload_databases/0`
+      `UAInspector.reload/0`
     - Supervision can now be done without starting the application
       ([#8](https://github.com/elixytics/ua_inspector/pulls/8))
     - The database downloader has been promoted to a directly usable module